Moved number of cores parameter for visualization from parameters to generalConfig. Added more csv outputs for visualization clustering

+++ datadriven/src/sgpp/datadriven/tools/hierarchyTree/ClusterNode.cpp
@@ -4,7 +4,7 @@
 // sgpp.sparsegrids.org
 
 #include  <sgpp/datadriven/tools/hierarchyTree/ClusterNode.hpp>
-
+#include <vector>
 #include <algorithm>
 #include <iostream>
 
@@ -20,13 +20,12 @@ ClusterNode::ClusterNode(int clusterLabel, std::vector<size_t> vertexIndexes, do
 }
 
 void ClusterNode::removeChild(ClusterNode* child) {
-  children.erase(std::remove(children.begin(), children.end(), child),children.end());
+  children.erase(std::remove(children.begin(), children.end(), child), children.end());
   delete child;
 }
 
 void ClusterNode::removeChildren() {
-
-  for (auto child: children) {
+  for (auto child : children) {
     delete child;
   }
   children.clear();
@@ -37,22 +36,16 @@ ClusterNode* ClusterNode::getParent() {
 }
 
 void ClusterNode::addChildren(std::vector<ClusterNode*> children) {
-  for (auto child: children) {
+  for (auto child : children) {
     this->addChild(child);
   }
 }
 
 bool ClusterNode::splitChild(ClusterNode* child, std::shared_ptr<Graph> graph,
   double densityThreshold) {
-
-  //std::cout << "Processing Node: " << this->getClusterLabel() << std::endl;
-  //std::cout << "Processing Child: " << child->getClusterLabel() << std::endl;
-
   std::vector<size_t> parentVertexIndexes = this->getVertexIndexes();
   std::vector<size_t> childVertexIndexes = child->getVertexIndexes();
 
-  //std::cout <<"Parent size: "<< parentVertexIndexes.size()<<std::endl;
-  //std::cout <<"Child size: "<< childVertexIndexes.size()<<std::endl;
   size_t maxConnectionsParent = (parentVertexIndexes.size()*(parentVertexIndexes.size()-1)/2);
   size_t maxConnectionsChildParent =
     childVertexIndexes.size() *(parentVertexIndexes.size() - childVertexIndexes.size());
@@ -61,7 +54,7 @@ bool ClusterNode::splitChild(ClusterNode* child, std::shared_ptr<Graph> graph,
   size_t connectionsChildParent = 0;
   std::vector<size_t> visitedVertex;
 
-  for(auto vertex: parentVertexIndexes) {
+  for (auto vertex : parentVertexIndexes) {
     if (graph->containsVertex(vertex)) {
       auto adjacentVertices = graph->getAdjacentVertices(vertex);
       for (auto adjacentVertex : adjacentVertices) {
@@ -74,15 +67,13 @@ bool ClusterNode::splitChild(ClusterNode* child, std::shared_ptr<Graph> graph,
             connectionsParent++;
           }
         }
-
       }
       visitedVertex.push_back(vertex);
     }
   }
- // std::cout << "Connections in parent " << connectionsParent << std::endl;
- // std::cout <<" Max connections parent " << maxConnectionsParent << std::endl;
+
   visitedVertex.clear();
-  for(auto vertex: childVertexIndexes) {
+  for (auto vertex : childVertexIndexes) {
     if (graph->containsVertex(vertex)) {
       auto adjacentVertices = graph->getAdjacentVertices(vertex);
       for (auto adjacentVertex : adjacentVertices) {
@@ -100,23 +91,18 @@ bool ClusterNode::splitChild(ClusterNode* child, std::shared_ptr<Graph> graph,
     }
   }
 
-  //std::cout << "Connections child parent " << connectionsChildParent << std::endl;
-  //std::cout <<" Max connections child parent " << maxConnectionsChildParent << std::endl;
-
-  double connectivityParent = connectionsParent/ static_cast<double>(maxConnectionsParent);
-  double connectivityChildParent = connectionsChildParent/static_cast<double>(maxConnectionsChildParent);
-
-  //std::cout << "Connectivty Parent" << connectivityParent << std::endl;
-  //std::cout << "Connectiviy Child Parent " << connectivityChildParent << std::endl;
+  double connectivityParent =
+    static_cast<double>(connectionsParent)/ static_cast<double>(maxConnectionsParent);
+  double connectivityChildParent =
+    static_cast<double>(connectionsChildParent)/static_cast<double>(maxConnectionsChildParent);
 
   double compare = connectivityChildParent/connectivityParent;
 
-  //std::cout << "Final value " << compare << std::endl;
   return compare <= densityThreshold;
 }
 
 bool ClusterNode::split(std::shared_ptr<Graph>  graph,  double densityThreshold) {
-  for(auto child:children) {
+  for (auto child : children) {
     if (splitChild(child, graph, densityThreshold)) {
       return true;
     }
@@ -161,6 +147,5 @@ size_t ClusterNode::getLevel() {
 void ClusterNode::setLevel(size_t level) {
   this->level = level;
 }
-
 }  // namespace datadriven
 }  // namespace sgpp
